requirements.txt
Django>=3.2
Pillow
opencv-python-headless
numpy
